requirements-test.txt
pytest>=8.0.0,<9.0.0
PyYAML>=6.0.0,<7.0.0
Jinja2>=3.1.0,<4.0.0